feat: move all helpers to util and merge test suites
